Added factory methods make_float, make_integer, and make_unsigned_integer.

+++ src/jsoncons/jsoncons_config.hpp
@@ -24,7 +24,7 @@
 
 namespace jsoncons {
 
-typedef double float_type;
+typedef long double float_type;
 typedef long long integer_type;
 typedef unsigned long long unsigned_integer_type;
